requirements.txt
hvac==2.3.0
pyyaml==6.0.1
kubernetes==29.0.0
click==8.1.2
